requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-filter==2.2.0
django-location-field==2.1.0
django-map-widgets==0.2.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
firebase-admin